Padronizing imagebox manipulation in expositions
