+++ StroopTest/Controllers/ExpositionController.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.Drawing;
     using System.Linq;
     using System.Windows.Forms;
     using TestPlatform.Views;
@@ -10,6 +11,49 @@
 
     public class ExpositionController
     {
+
+        private static int brush25 = 25;
+        private static int brush4 = 4;
+
+        public static void makingFixPoint(string fixPoint, string fixPointColor, Form expositionForm)
+        {
+            SolidBrush myBrush = new SolidBrush(ColorTranslator.FromHtml(fixPointColor));
+            float[] clientMiddle = { (expositionForm.ClientSize.Width / 2), (expositionForm.ClientSize.Height / 2) };
+
+            switch (fixPoint)
+            {
+                case "+": // cross fixPoint
+                    float widthCross = 2 * brush25;
+                    float heightCross = 2 * brush4;
+                    Graphics formGraphicsCross1 = expositionForm.CreateGraphics();
+                    Graphics formGraphicsCross2 = expositionForm.CreateGraphics();
+                    float xCross1 = clientMiddle[0] - (widthCross / 2);
+                    float yCross1 = clientMiddle[1] - (heightCross / 2);
+                    float xCross2 = clientMiddle[0] - (heightCross / 2);
+                    float yCross2 = clientMiddle[1] - (widthCross / 2);
+
+                    formGraphicsCross1.FillRectangle(myBrush, xCross1, yCross1, widthCross, heightCross);
+                    formGraphicsCross2.FillRectangle(myBrush, xCross2, yCross2, heightCross, widthCross);
+                    formGraphicsCross1.Dispose();
+                    formGraphicsCross2.Dispose();
+                    break;
+                case "o": // circle fixPoint
+                    Graphics formGraphicsEllipse = expositionForm.CreateGraphics();
+                    float xEllipse = clientMiddle[0];
+                    float yEllipse = clientMiddle[1];
+                    float widthEllipse = 2 * brush25;
+                    float heightEllipse = 2 * brush25;
+                    formGraphicsEllipse.FillEllipse(myBrush, xEllipse, yEllipse, widthEllipse, heightEllipse);
+                    formGraphicsEllipse.Dispose();
+                    break;
+                default:
+                    break;
+
+            }
+            myBrush.Dispose();
+        }
+
+
         public static string[] ShuffleArray(string[] array, int expectedLength, int rndSeed) // randomiza Vetor - parâmetros: vetor / tamanho esperado do vetor randomizado
         {
             List<string> randomArray = new List<string>();
@@ -113,6 +157,15 @@
             {
                 MessageBox.Show(ex.Message);
             }
+        }
+
+        public static PictureBox InitializeImageBox(int stimuliSize, Image image)
+        {
+            PictureBox newPictureBox = new PictureBox();
+            newPictureBox.SizeMode = PictureBoxSizeMode.StretchImage;
+            newPictureBox.Size = new Size(stimuliSize, stimuliSize);
+            newPictureBox.Image = image;
+            return newPictureBox;
         }
 
         public static void BeginExperimentTest(string programName, string participantName, char mark, Form form)
